requirements.txt
python-dotenv==1.1.0
streamlit==1.45.1
nats-py==2.10.0
openai==1.84.0